Move current_user viewset into users file
