+++ galaxy_ng/app/api/ui/serializers/__init__.py
@@ -16,12 +16,9 @@ from .namespace import (
     NamespaceUpdateSerializer
 )
 
-from .current_user import (
-    CurrentUserSerializer
-)
-
 from .user import (
-    UserSerializer
+    UserSerializer,
+    CurrentUserSerializer
 )
