--- galaxy_ng/app/api/ui/serializers/current_user.py
@@ -1,5 +1,0 @@
-from rest_framework import serializers
-
-
-class CurrentUserSerializer(serializers.Serializer):
-    is_partner_engineer = serializers.BooleanField()
