+++ galaxy_ng/app/api/ui/viewsets/__init__.py
@@ -1,8 +1,7 @@
 from .namespace import NamespaceViewSet, MyNamespaceViewSet
 from .collection import CollectionViewSet, CollectionVersionViewSet, CollectionImportViewSet
 from .tags import TagsViewSet
-from .current_user import CurrentUserViewSet
-from .user import UserViewSet
+from .user import UserViewSet, CurrentUserViewSet
 
 __all__ = (
     'NamespaceViewSet',
