+++ galaxy_ng/app/api/ui/viewsets/user.py
@@ -1,4 +1,5 @@
 from rest_framework import mixins
+from rest_framework.response import Response
 
 from galaxy_ng.app.models import auth as auth_models
 from galaxy_ng.app.api import permissions
@@ -20,3 +21,14 @@ class UserViewSet(
 
     def get_queryset(self):
         return auth_models.User.objects.all()
+
+
+class CurrentUserViewSet(api_base.GenericViewSet):
+    serializer_class = serializers.CurrentUserSerializer
+
+    def retrieve(self, request, *args, **kwargs):
+        data = serializers.CurrentUserSerializer({
+            'is_partner_engineer': permissions.IsPartnerEngineer().has_permission(request, self)
+        }).data
+
+        return Response(data)
